topic_11_Python_RAG/docsum/requirements.txt
groq
beautifulsoup4
lxml
python-dotenv
nltk